requirements.txt
fastapi
uvicorn[standard]
routeros_api
pyjwt
sqlalchemy